@@ -21,7 +21,7 @@
 project(STP)
 cmake_minimum_required(VERSION 3.3 FATAL_ERROR)
 set(CMAKE_C_STANDARD 99)
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 14)
 
 include (GenerateExportHeader)
 include(GNUInstallDirs)
